requirements.txt
earthengine-api==0.1.405
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
google-auth==2.35.0
